Перенес вспомогательные функции из const.js в util.js
